WIP: Encode bytes as mp_bin on Python 3
This allows to write mp_bin, which is required for 'varbinary' field
type: just write a value of 'bytes' type and it will be encoded as
mp_bin instead of mp_str.

mp_bin is already decoded into bytes, which is consistent with the new
encode behaviour.

XXX: Change the behaviour only under an option? If we'll go this way, we
can also split mp_str / mp_bin across unicode / str for Python 2, but
don't sure it is convenient.

Fixes #105